Allow configuration of instance port range
This helps when defining a firewall range to open up access to the
Postgres instances.

+++ pkg/models/instance.go
@@ -12,7 +12,7 @@ type Instance struct {
 	RefreshToken string
 	CreatedAt    time.Time `jsonapi:"attr,created_at,iso8601"`
 	UpdatedAt    time.Time `jsonapi:"attr,updated_at,iso8601"`
-	Port         int       `jsonapi:"attr,port"`
+	Port         uint16    `jsonapi:"attr,port"`
 
 	Credentials *InstanceCredentials `jsonapi:"relation,credentials"`
 }
